+++ src/Auth/CachedEloquentUserProvider.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace DiegoVasconcelos\AuthCache\Auth;
 
+use DiegoVasconcelos\AuthCache\Contracts\Cache\CacheInterface;
+use DiegoVasconcelos\AuthCache\Contracts\Cache\CacheKeyGeneratorInterface;
 use DiegoVasconcelos\AuthCache\DTOs\CachedUserData;
 use Illuminate\Auth\EloquentUserProvider;
 use Illuminate\Contracts\Auth\Authenticatable;
@@ -14,16 +16,17 @@ class CachedEloquentUserProvider extends EloquentUserProvider
     public function __construct(
         $hasher,
         $model,
-        private CacheManager $cacheManager,
+        private CacheInterface $cache,
+        private CacheKeyGeneratorInterface $keyGenerator,
     ) {
         parent::__construct($hasher, $model);
     }
 
     public function retrieveById($identifier): string|(Model&Authenticatable)|null
     {
-        $cachedData = $this->cacheManager->remember(
-            key: $this->cacheManager->generateKey($this->getModel(), $identifier),
-            ttl: now()->addMinutes($this->cacheManager->getTtl()),
+        $cachedData = $this->cache->remember(
+            key: $this->keyGenerator->generate($this->getModel(), $identifier),
+            ttl: now()->addMinutes($this->cache->getTtl()),
             callback: function () use ($identifier) {
                 $result = parent::retrieveById($identifier);